Moved simple typealias targets back into main macro test target
